+++ src/apps/portal/src/main/java/com/opentravelsoft/action/sms/SendAction.java
@@ -42,7 +42,7 @@ public class SendAction extends PortalAction {
     String pswd = getConfig("ebiz_sms_pswd");
     String extno = getConfig("ebiz_sms_extno");
 
-    long userId = 0;
+    int userId = 0;
     // uid=xkh&pswd=5537&msg=+_()*^%$@!hello%20world&extno=5537&mob=13816617311
 
     if (user != null) {
