+++ src/modules/hibernate/src/main/java/com/opentravelsoft/providers/TeamDao.java
@@ -8,7 +8,7 @@ import com.opentravelsoft.util.PaginationSupport;
 
 public interface TeamDao extends GenericDao<Team, Integer> {
 
-  public List<Team> getTeam(long userId, TeamType type);
+  public List<Team> getTeam(int userId, TeamType type);
 
   public PaginationSupport getTeamList(int fromRecord, int pageSize);
