+++ src/modules/hibernate/src/main/java/com/opentravelsoft/providers/product/ProductItemDao.java
@@ -6,12 +6,11 @@ import com.opentravelsoft.entity.ProductItem;
 import com.opentravelsoft.entity.product.ProductType;
 import com.opentravelsoft.providers.GenericDao;
 
-public interface ProductItemDao extends GenericDao<ProductItem, Long>
-{
-    List<ProductItem> getProductItems(boolean tmpl);
+public interface ProductItemDao extends GenericDao<ProductItem, Integer> {
+  List<ProductItem> getProductItems(boolean tmpl);
 
-    ProductItem getProdutItem(String itemCode);
+  ProductItem getProdutItem(String itemCode);
 
-    List<ProductItem> getProductItems(ProductType type);
+  List<ProductItem> getProductItems(ProductType type);
 
 }
