+++ src/modules/service/src/main/java/com/opentravelsoft/service/SmsService.java
@@ -2,14 +2,13 @@ package com.opentravelsoft.service;
 
 import java.util.Date;
 
-public interface SmsService
-{
-    public int txReceive(String seqno, String mob, String msg, String srcMsg);
+public interface SmsService {
+  public int txReceive(String seqno, String mob, String msg, String srcMsg);
 
-    public int txSend(String seqno, String msg, String mob, long userId);
+  public int txSend(String seqno, String msg, String mob, int userId);
 
-    public int txSendResult(String seqno, String extno, String mob,
-            String stat, String status);
+  public int txSendResult(String seqno, String extno, String mob, String stat,
+      String status);
 
-    public Date roGetSysdate();
+  public Date roGetSysdate();
 }
