+++ src/modules/service/src/main/java/com/opentravelsoft/service/order/PlanSearchService.java
@@ -8,10 +8,10 @@ import com.opentravelsoft.entity.Plan;
 import com.opentravelsoft.entity.Team;
 
 public interface PlanSearchService {
-  public List<Plan> roFind(String lineName, long teamId, long userId,
+  public List<Plan> roFind(String lineName, int teamId, int userId,
       Date startDate, Date endDate, double lowerPrice, double upperPrice,
       boolean openKey);
 
-  public List<Team> getTeamList(long userId, TeamType type);
+  public List<Team> getTeamList(int userId, TeamType type);
 
 }
